+++ NEWS.md
@@ -4,6 +4,9 @@
 - Dynamically set a node in a Zarr hierarchy.
 - Set metadata on a node in a memory store.
 - Convention classes are now coded as attribute factories.
+- Zarr package options can now be retrieved and modified with the `zarr_options()` function.
+- Compute optimal chunking sizes from array shape when not set explicitly.
+- In `as_zarr()`, small arrays are not compressed. This is controlled by the `Zarr.options$min_compress` setting.
 - Fix key listing in memory stores.
 
 # zarr 0.4.1